Evm staking implement. (#745)
* bump chaindev (#757)


* update EVM_FIRST_BLOCK_HEIGHT to checkpoint (#743)

* Gas calculation of EIP-1962 (#759)

* added gas calculation and updated unit test

* Empty-Commit

* Apply suggestions from code review


* Fix errors after clippy updatd (#766)

* Block confidential transfer with tracing policies (#715)

* add the blocker

* fmt

* fmt

* add the restriction

* fmt

* Refactored code for asset tracing condition check

* Removed debug lines


* Use Rust optimization when compiling the instance for Nightly test (#767)

* Update Dockerfile-binary-image-dev

* Empty-Commit


* Revert "Use Rust optimization when compiling the instance for Nightly test (#767)" (#774)

This reverts commit 5b0340672aec398d7a4798c21589fb9b102ca4be.

* add wasm getkeyold (#778)

* add wasm metamask account (#779)

* fix build macos fn (#770)

* add wasm get_keypair_by_pri_key (#782)

* Implement call stake, delegate, undelegate, claim.

* add wasm bech32_to_base64_old (#784)


* add wasm new_keypair_old (#789)


* implement mint from evm staking.

* Upgrade prism (#791)

* Mint FRA for test address

* upgrade prism

* make fmt

* make test

* fix make test

* fix withdraw_fra and parse_deposit_asset

* clean code.

* implement mint from claim.

* remove get claim ops, use event to mint.

---------

+++ src/components/contracts/primitives/types/src/crypto.rs
@@ -450,11 +450,12 @@ impl Verify for XfrSignature {
 pub fn secp256k1_ecdsa_recover(sig: &[u8; 65], msg: &[u8; 32]) -> ruc::Result<[u8; 64]> {
     let rs = libsecp256k1::Signature::parse_standard_slice(&sig[0..64])
         .map_err(|_| eg!("Ecdsa signature verify error: bad RS"))?;
-    let v =
-        libsecp256k1::RecoveryId::parse(
-            if sig[64] > 26 { sig[64] - 27 } else { sig[64] } as u8,
-        )
-        .map_err(|_| eg!("Ecdsa signature verify error: bad V"))?;
+    let v = libsecp256k1::RecoveryId::parse(if sig[64] > 26 {
+        sig[64] - 27
+    } else {
+        sig[64]
+    })
+    .map_err(|_| eg!("Ecdsa signature verify error: bad V"))?;
     let pubkey = libsecp256k1::recover(&libsecp256k1::Message::parse(msg), &rs, &v)
         .map_err(|_| eg!("Ecdsa signature verify error: bad signature"))?;
     let mut res = [0u8; 64];
